return fee estimates for block from 1 to 25
first element is min relay fee

+++ subprojects/gdk_rust/gdk_common/src/lib.rs
@@ -68,9 +68,8 @@ pub struct TxListItem {
 }
 
 // This one is simple enough to derive a serializer
-#[serde(transparent)] // serialize as f64
 #[derive(Serialize, Debug)]
-pub struct FeeEstimate(pub f64);
+pub struct FeeEstimate(pub u64);
 
 pub struct AddressResult(pub String);
 pub struct TxsResult(pub Vec<TxListItem>);
